refactor(logger): split the 618-line module along its four seams
- trace.ts / wide-event.ts / environment.ts / otlp.ts with index.ts as the package-root entry
- Delete the annotateWide identity alias; callers use Effect.annotateLogsScoped directly
- Remove the duplicated verbatim sink-dispatch comment; document env-var precedence in WideEventEnvironment
