+++ ARCHITECTURE.md
@@ -67,7 +67,7 @@ React Email templates and Cloudflare Email Service sending boundary. Outbound em
 - **Rate limiting:** Cloudflare's native `RateLimit` binding (one per bucket, specs in [`infra/bindings.ts`](./infra/bindings.ts), attached by alchemy and mirrored in wrangler under the drift test). The runtime layer ([`apps/api/src/rate-limit.ts`](./apps/api/src/rate-limit.ts), [`apps/web/src/lib/rate-limit.ts`](./apps/web/src/lib/rate-limit.ts)) calls `binding.limit({ key })` when the binding is present and falls back to a per-isolate `HashMap` brake when it isn't (local dev/tests). Distributed limits are eventually-consistent across regions — swap to a Durable Object if you need strong consistency.
 - **Webhook delivery:** The background worker's `queue` handler processes batches with `Promise.all`, calls `message.retry({ delaySeconds })` on failure (no manual `WEBHOOK_QUEUE.send` requeue), and Cloudflare delivers to `b2b-saas-starter-webhooks-dlq` after `maxRetries` (6).
 - **Observability:** one seam, [`packages/logger`](./packages/logger/src/index.ts), covering wide events, traces, and metrics (ADR 0050).
-  - **Wide events.** One canonical line per request per service. `withRequestScope` opens a span plus a `Scope`, seeds annotations with `service`/`traceId`/`otelTraceId`/`otelSpanId`/environment/metadata, and emits the event from `Effect.onExit` — not a scope finalizer, which runs after `annotateLogsScoped` has already restored the previous annotations and would silently drop everything handlers added. `annotateWide` (alias of `Effect.annotateLogsScoped`) is how handlers add business context to that same event. The level is `info` on success and `error` on failure, with the `Cause` attached; there is no third level. `readWideEventEnvironment` pulls `commitHash`/`serviceVersion`/`region`/`environment` from worker env + cf hints. Workers never assemble the envelope by hand: `withHttpRequestScope({ service, event, request, env })` owns the HTTP recipe and `withTriggerScope` the queue one, so a format change edits one file.
+  - **Wide events.** One canonical line per request per service. `withRequestScope` opens a span plus a `Scope`, seeds annotations with `service`/`traceId`/`otelTraceId`/`otelSpanId`/environment/metadata, and emits the event from `Effect.onExit` — not a scope finalizer, which runs after `annotateLogsScoped` has already restored the previous annotations and would silently drop everything handlers added. `Effect.annotateLogsScoped` (alias of `Effect.annotateLogsScoped`) is how handlers add business context to that same event. The level is `info` on success and `error` on failure, with the `Cause` attached; there is no third level. `readWideEventEnvironment` pulls `commitHash`/`serviceVersion`/`region`/`environment` from worker env + cf hints. Workers never assemble the envelope by hand: `withHttpRequestScope({ service, event, request, env })` owns the HTTP recipe and `withTriggerScope` the queue one, so a format change edits one file.
   - **Traces.** W3C `traceparent` (and B3) in and out, via `HttpTraceContext`. Inbound headers become the request span's parent; Effect's `HttpClient` injects the headers on every outbound call and spans D1 queries on the way through, so a page load is one trace from the browser request down to `sql.execute`. The queue is bridged explicitly: `WebhookPublisher` stamps `currentTraceparent` onto the message and the background consumer continues that trace instead of starting a new one. `x-trace-id` survives as the human-quotable correlation key and now defaults to the OTel trace id, so one value resolves in both the log stream and the trace backend.
   - **Metrics.** `starter.requests` (counter) and `starter.request.duration` (histogram) come off the same scope that emits the wide event, attributed by `service`/`event`/`status` only — high-cardinality dimensions belong on the event and the span.
   - **Export.** `makeOtlpLayer(service, env)` exports all three signals over OTLP/HTTP when `OTEL_EXPORTER_OTLP_ENDPOINT` is set, and is `Layer.empty` when it isn't. It is provided per invocation, never per isolate — see ADR 0050. `WideEventLoggerLive` (console JSON + the tracer logger) stays isolate-level in all three workers. `apps/web` opens exactly one such scope per request in a global TanStack Start request middleware ([`src/start.ts`](./apps/web/src/start.ts)); loaders and server functions join it through `withWebRequestScope` rather than each opening their own.
