+++ apps/web/AGENTS.md
@@ -33,7 +33,7 @@ TanStack Start web app served by a Cloudflare Worker (`vite.config.ts` uses the 
 ## Observability
 
 - **One wide event per HTTP request, opened in exactly one place.** `src/start.ts` registers a global TanStack Start request middleware (`createStart({ requestMiddleware })`) that wraps every server request — SSR document renders and server-function calls — in `runWebRequestScope` (`src/lib/observability.ts`). It continues any inbound `traceparent`, provides `makeOtlpLayer` per invocation (never per isolate — see ADR 0050), and emits the canonical line with `statusCode` and a `nested` array when the response is ready. `WideEventLoggerLive` is provided isolate-level from a module-scope `ManagedRuntime`, not rebuilt per request.
-- **Nested work joins that scope; it never opens its own.** `runWorkspaceCapabilities` / `runCapabilities`, the `/api/auth/$` handler, and the auth gates in `src/lib/server/auth.ts` / `src/lib/server/admin.ts` all call `withWebRequestScope({ event, metadata }, effect)`, which becomes a child span under the request span, reuses the request's loggers and exporters, and folds its own `annotateWide` fields into the request's event. Without it a page load emits one wide event and one root trace _per loader_. Don't call `withTriggerScope` directly from a loader.
+- **Nested work joins that scope; it never opens its own.** `runWorkspaceCapabilities` / `runCapabilities`, the `/api/auth/$` handler, and the auth gates in `src/lib/server/auth.ts` / `src/lib/server/admin.ts` all call `withWebRequestScope({ event, metadata }, effect)`, which becomes a child span under the request span, reuses the request's loggers and exporters, and folds its own `Effect.annotateLogsScoped` fields into the request's event. Without it a page load emits one wide event and one root trace _per loader_. Don't call `withTriggerScope` directly from a loader.
 - Outside a request — a client-side navigation re-running loaders against the Seed layer, or a standalone test — `withWebRequestScope` falls back to its own scope and tags the event `scope: 'standalone'`, so a missed join is visible in the log stream instead of silently producing an extra root trace.
 - The registry is a `WeakMap` keyed by the request object (an isolate interleaves concurrent requests, so a module-level "current" would race). `src/lib/request-context.ts` exposes the lookup as `currentRequest()`, wrapping `createIsomorphicFn`/`getRequest` in one place: it keeps `getRequest` out of the browser bundle, returns `undefined` rather than throwing with no ambient request, and is the seam tests mock.
 - `authRuntime` holds only the `Auth` service. Loggers, tracer, and exporters belong to one request and must not live on a per-isolate runtime — the auth gates reach them by joining the request scope (above), not by reading `authRuntime`.
